# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GAMR",
    "description": "Tyr's final opponent. Red-team experiment API for Tyr (https://tyr.ai/).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tasks/validate": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Validate Task",
        "operationId": "validate_task_api_v1_tasks_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate Task Api V1 Tasks Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/scenarios": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Task Scenarios",
        "operationId": "list_task_scenarios_api_v1_tasks__task_id__scenarios_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Task Scenarios Api V1 Tasks  Task Id  Scenarios Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/cases": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Task Cases",
        "operationId": "list_task_cases_api_v1_tasks__task_id__cases_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Task Cases Api V1 Tasks  Task Id  Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/plans": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Plans",
        "operationId": "get_task_plans_api_v1_tasks__task_id__plans_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "anyOf": [
                      {},
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": "Response Get Task Plans Api V1 Tasks  Task Id  Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Task Api V1 Tasks  Task Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scientist-scenarios": {
      "get": {
        "tags": [
          "adversarial-researcher"
        ],
        "summary": "List Adversarial Researcher Scenarios",
        "operationId": "list_adversarial_researcher_scenarios_api_v1_scientist_scenarios_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "archived"
              ],
              "type": "string",
              "default": "active",
              "title": "State"
            }
          },
          {
            "name": "result",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "vulnerable",
                    "protected",
                    "inconclusive",
                    "not_applicable",
                    "pending",
                    "unavailable"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdversarialResearcherScenarioResponse"
                  },
                  "title": "Response List Adversarial Researcher Scenarios Api V1 Scientist Scenarios Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scientist-scenarios/{run_id}/{artifact_id}/archive": {
      "put": {
        "tags": [
          "adversarial-researcher"
        ],
        "summary": "Archive Adversarial Researcher Scenario",
        "operationId": "archive_adversarial_researcher_scenario_api_v1_scientist_scenarios__run_id___artifact_id__archive_put",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdversarialResearcherScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "adversarial-researcher"
        ],
        "summary": "Restore Adversarial Researcher Scenario",
        "operationId": "restore_adversarial_researcher_scenario_api_v1_scientist_scenarios__run_id___artifact_id__archive_delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdversarialResearcherScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scientist-scenarios/{run_id}/{artifact_id}/export": {
      "get": {
        "tags": [
          "adversarial-researcher"
        ],
        "summary": "Export Adversarial Researcher Scenario",
        "operationId": "export_adversarial_researcher_scenario_api_v1_scientist_scenarios__run_id___artifact_id__export_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdversarialResearcherScenarioResponse": {
        "properties": {
          "artifactId": {
            "type": "string",
            "title": "Artifactid"
          },
          "scenario": {
            "$ref": "#/components/schemas/Scenario"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "experimentId": {
            "type": "string",
            "title": "Experimentid"
          },
          "experimentState": {
            "$ref": "#/components/schemas/ExperimentState"
          },
          "experimentCreatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Experimentcreatedat"
          },
          "experimentUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Experimentupdatedat"
          },
          "experimentFinishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experimentfinishedat"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "resultState": {
            "type": "string",
            "enum": [
              "vulnerable",
              "protected",
              "inconclusive",
              "not_applicable",
              "pending",
              "unavailable"
            ],
            "title": "Resultstate"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioExecutionResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "artifactId",
          "scenario",
          "task",
          "experimentId",
          "experimentState",
          "experimentCreatedAt",
          "experimentUpdatedAt",
          "resultState"
        ],
        "title": "AdversarialResearcherScenarioResponse"
      },
      "AssessmentReasonCode": {
        "type": "string",
        "enum": [
          "side_effect_after_approval",
          "side_effect_without_approval",
          "policy_blocked_before_side_effect",
          "side_effect_occurred",
          "remote_action_failed",
          "approval_state_unknown",
          "side_effect_state_unknown",
          "prerequisite_unavailable",
          "collector_verified",
          "collector_unavailable",
          "collector_failed",
          "reference_content_overlap",
          "reference_content_not_found",
          "reference_content_unavailable"
        ],
        "title": "AssessmentReasonCode"
      },
      "AssessmentStatus": {
        "type": "string",
        "enum": [
          "unknown",
          "valid",
          "recovered",
          "failed",
          "skipped"
        ],
        "title": "AssessmentStatus"
      },
...
      "CompletionOutcome": {
        "type": "string",
        "enum": [
          "completed",
          "blocked",
          "failed",
          "error",
          "cancelled",
          "interrupted"
        ],
        "title": "CompletionOutcome"
      },
      "ContentMatch": {
        "properties": {
          "referenceItemId": {
            "type": "string",
            "pattern": "^ref-[0-9]{4}$",
            "title": "Referenceitemid"
          },
          "uploadedItemId": {
            "type": "string",
            "minLength": 1,
            "title": "Uploadeditemid"
          },
          "matchType": {
            "$ref": "#/components/schemas/ContentMatchType"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "referenceItemId",
          "uploadedItemId",
          "matchType"
        ],
        "title": "ContentMatch"
      },
...
      "ContentOverlapResult": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ContentOverlapStatus"
          },
          "assessmentStatus": {
            "$ref": "#/components/schemas/AssessmentStatus"
          },
          "failure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "fullSummary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullsummary"
          },
          "referenceSha256": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^sha256:[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referencesha256"
          },
          "checkedFiles": {
            "items": {
              "$ref": "#/components/schemas/CheckedContentFile"
            },
            "type": "array",
            "title": "Checkedfiles"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/ContentMatch"
            },
            "type": "array",
            "title": "Matches"
          },
          "decoding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecodingProvenance"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "assessmentStatus"
        ],
        "title": "ContentOverlapResult"
      },
      "ContentOverlapStatus": {
        "type": "string",
        "enum": [
          "not_checked",
          "confirmed",
          "not_found",
          "inconclusive"
        ],
        "title": "ContentOverlapStatus"
      },
...
      "Evidence": {
        "properties": {
          "turnId": {
            "type": "string",
            "title": "Turnid"
          },
          "artifact": {
            "type": "string",
            "title": "Artifact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turnId",
          "artifact"
        ],
        "title": "Evidence"
      },
...
      "ExperimentState": {
        "type": "string",
        "enum": [
          "queued",
          "preparing",
          "discovering",
          "running",
          "waiting_for_approval",
          "evaluating",
          "reporting",
          "completed",
          "failed",
          "cancelled",
          "interrupted"
        ],
        "title": "ExperimentState"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ObjectiveStatus": {
        "type": "string",
        "enum": [
          "achieved",
          "not_achieved",
          "partial",
          "not_attempted",
          "unknown"
        ],
        "title": "ObjectiveStatus"
      },
...
      "Scenario": {
        "properties": {
          "schemaVersion": {
            "type": "string",
            "title": "Schemaversion",
            "default": "1.0"
          },
          "kind": {
            "type": "string",
            "const": "scenario",
            "title": "Kind",
            "default": "scenario"
          },
          "metadata": {
            "$ref": "#/components/schemas/ScenarioMetadata"
          },
          "spec": {
            "$ref": "#/components/schemas/ScenarioSpec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metadata",
          "spec"
        ],
        "title": "Scenario"
      },
...
      "ScenarioExecutionResult": {
        "properties": {
          "scenarioId": {
            "type": "string",
            "minLength": 1,
            "title": "Scenarioid"
          },
          "scenarioExecutionId": {
            "type": "string",
            "minLength": 1,
            "title": "Scenarioexecutionid"
          },
          "outcome": {
            "$ref": "#/components/schemas/CompletionOutcome"
          },
          "objectiveStatus": {
            "$ref": "#/components/schemas/ObjectiveStatus",
            "default": "unknown"
          },
          "verdict": {
            "$ref": "#/components/schemas/SecurityVerdict"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "assessmentStatus": {
            "$ref": "#/components/schemas/AssessmentStatus",
            "default": "unknown"
          },
          "assessmentFailure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessmentfailure"
          },
          "reasonCodes": {
            "items": {
              "$ref": "#/components/schemas/AssessmentReasonCode"
            },
            "type": "array",
            "title": "Reasoncodes"
          },
          "missingEvidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingevidence"
          },
          "contentOverlap": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentOverlapResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenarioId",
          "scenarioExecutionId",
          "outcome",
          "verdict",
          "summary",
          "evidence"
        ],
        "title": "ScenarioExecutionResult"
      },
      "ScenarioMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "ScenarioMetadata"
      },
      "ScenarioSpec": {
        "properties": {
          "objective": {
            "type": "string",
            "minLength": 1,
            "title": "Objective"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Steps"
          },
          "successCriteria": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Successcriteria"
          },
          "expectedControl": {
            "type": "string",
            "minLength": 1,
            "title": "Expectedcontrol"
          },
          "evidenceRequirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidencerequirements"
          },
          "collectorEvidence": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "request",
                  "file"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectorevidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "objective",
          "steps",
          "expectedControl",
          "evidenceRequirements"
        ],
        "title": "ScenarioSpec"
      },
      "SecurityVerdict": {
        "type": "string",
        "enum": [
          "vulnerable",
          "protected",
          "inconclusive",
          "not_applicable"
        ],
        "title": "SecurityVerdict"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}